==> notebooks/NB01-Data-Collection-Score-Info.ipynb <==
from scrapy import Selector

import requests

from scrapy import Selector

import pandas as pd

from tqdm.notebook import tqdm
tqdm.pandas()

# Loads dataframe which contains all games
all_games = pd.read_csv('../data/raw data/all_games.csv')

# Remove the ', [, and ] characters from the 'link' column in 'all_games.csv' so that scrapy can access the links
all_games['link'] = all_games['link'].str.replace('[', '')
all_games['link'] = all_games['link'].str.replace(']', '')
all_games['link'] = all_games['link'].str.replace("'", '')

# Function to scrape the score information of a game
def get_score_info(link):
    response = requests.get(link)

    sel = Selector(text = response.text)

    containers = sel.css('div.info-score')

    game_title = Selector(text = response.text).css('div.mb h1::text').get() 

    moby_score = containers.css('.mobyscore::text').get()

    critics_score = containers.css('dl dd:nth-child(4)::text').get()

    player_score = containers.css('.stars-sm::attr(data-tooltip)').get()

    if 'Xbox' in containers.css('#platformRankings a::text').getall(): # Sometimes the Xbox score is not available, this if loop sets the value to 'N/A' if it is not available
        key = containers.css('#platformRankings a::text').getall()

        values = containers.css('#platformRankings span::text').getall()

        ranking = dict(zip(key, values))

    else:
        ranking = {'Xbox': 'N/A'}


    
    return{
            'title': game_title,
            'moby_score': moby_score,
            'critics_score': critics_score,
            'player_score': player_score,
            'xbox_score': ranking['Xbox']   
        }

# Applies the function to all the games in the dataframe
score_dict = all_games['link'].progress_apply(get_score_info)

# Turn score_dict into a dataframe
score = pd.DataFrame(score_dict.tolist())

# Save the dataframe to a csv file
score.to_csv('../data/raw data/score_info.csv', index = False)
