==> notebooks/NB03-Database-Creation.ipynb <==
import sqlite3

import pandas as pd

# Loading in all the data frames
all_games = pd.read_csv('../data/clean data/all_games_clean.csv')
genre_info = pd.read_csv('../data/clean data/genre_info_clean.csv')
score_info = pd.read_csv('../data/clean data/score_info_clean.csv')
release_info = pd.read_csv('../data/clean data/release_info_clean.csv')

conn = sqlite3.connect('../Data/MobyGames.db')

# Laod all_games into the database I created which is called MobyGames.db
all_games.to_sql('all_games', conn, if_exists='replace', index=False)

# Load genre_info into the same database but as a new table
genre_info.to_sql('genre_info', conn, if_exists='replace', index=False)

# Load score_info into the same database but as a new table
score_info.to_sql('score_info', conn, if_exists='replace', index=False)

# Load release_info into the same database but as a new table
release_info.to_sql('release_info', conn, if_exists='replace', index=False)

conn.close()
